Move PublicKey and Address classes to models (#66)
