+++ wallet-kit/src/androidTest/java/bitcoin/wallet/kit/transactions/TransactionLinkerTest.kt
@@ -2,7 +2,7 @@ package bitcoin.wallet.kit.transactions
 
 import bitcoin.wallet.kit.RealmFactoryMock
 import bitcoin.wallet.kit.core.toHexString
-import bitcoin.wallet.kit.hdwallet.PublicKey
+import bitcoin.wallet.kit.models.PublicKey
 import bitcoin.wallet.kit.models.Transaction
 import bitcoin.wallet.kit.models.TransactionInput
 import bitcoin.wallet.kit.models.TransactionOutput
