+++ wallet-kit/src/test/kotlin/bitcoin/wallet/kit/managers/BlockDiscoverTest.kt
@@ -1,7 +1,7 @@
 package bitcoin.wallet.kit.managers
 
 import bitcoin.wallet.kit.core.hexStringToByteArray
-import bitcoin.wallet.kit.hdwallet.Address
+import bitcoin.wallet.kit.models.Address
 import bitcoin.wallet.kit.models.Block
 import bitcoin.wallet.kit.network.NetworkParameters
 import bitcoin.wallet.kit.utils.AddressConverter
